TP/AI_projeto/Projeto_IA_14/requirements.txt
ultralytics
opencv-python
numpy
pyautogui
mss
customtkinter
python-chess
pygame
cairosvg